requirements.txt
python-telegram-bot>=20.0
pytz
python-dotenv
rapidfuzz